+++ tests/test_track.py
@@ -77,6 +77,21 @@ class TrackTest(unittest.TestCase):
 
         self.assertEqual(result, 'Track(%r)' % 'foo')
 
+    def test_eq(self, lib_mock):
+        sp_track = spotify.ffi.cast('sp_track *', 42)
+        track1 = spotify.Track(self.session, sp_track=sp_track)
+        track2 = spotify.Track(self.session, sp_track=sp_track)
+
+        self.assertTrue(track1 == track2)
+        self.assertFalse(track1 == 'foo')
+
+    def test_ne(self, lib_mock):
+        sp_track = spotify.ffi.cast('sp_track *', 42)
+        track1 = spotify.Track(self.session, sp_track=sp_track)
+        track2 = spotify.Track(self.session, sp_track=sp_track)
+
+        self.assertFalse(track1 != track2)
+
     def test_is_loaded(self, lib_mock):
         lib_mock.sp_track_is_loaded.return_value = 1
         sp_track = spotify.ffi.cast('sp_track *', 42)